Layout_TestCase/http/tests/xmlhttprequest/onerror-event-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/access-control-basic-denied.cgi. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
This test that the error event is fired for XMLHttpRequests

PASS: error event fired.
